% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeachingMaterials.r
\docType{data}
\name{covid_host}
\alias{covid_host}
\title{Covid host data}
\format{
factor
}
\usage{
data(covid_host)
}
\description{
Data on Covid hosts from Yang et al 2023
}
\examples{
data(covid_host)
head(covid_host)
}
\references{
Yang J, Skaro M, Chen J, Zhan D, Lyu L, Gay S, Kandeil A, Ali MA, Kayali G, Stoianova K, Ji P. The species coalescent indicates possible bat and pangolin origins of the COVID-19 pandemic. Scientific Reports. 2023 Apr 5;13(1):5571.
}
\keyword{datasets}
