% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeachingMaterials.r
\docType{data}
\name{covid_tree}
\alias{covid_tree}
\title{Covid species tree}
\format{
phylo
}
\usage{
data(covid_tree)
}
\description{
Subsampled covid species tree from Yang et al 2023
}
\examples{
data(covid_tree)
plot(covid_tree, cex = 0.5)
}
\references{
Yang J, Skaro M, Chen J, Zhan D, Lyu L, Gay S, Kandeil A, Ali MA, Kayali G, Stoianova K, Ji P. The species coalescent indicates possible bat and pangolin origins of the COVID-19 pandemic. Scientific Reports. 2023 Apr 5;13(1):5571.
}
\keyword{datasets}
