% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeachingMaterials.r
\docType{data}
\name{H5N1_host}
\alias{H5N1_host}
\title{H5N1 host data}
\format{
factor
}
\usage{
data(H5N1_host)
}
\description{
H5N1 host data from Nguyen et al 2024
}
\examples{
data(H5N1_host)
head(H5N1_host)
}
\references{
Nguyen TQ, Hutter C, Markin A, Thomas MN, Lantz K, Killian ML, Janzen GM, Vijendran S, Wagle S, Inderski B, Magstadt DR. Emergence and interstate spread of highly pathogenic avian influenza A (H5N1) in dairy cattle. bioRxiv. 2024:2024-05.
}
\keyword{datasets}
