% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeachingMaterials.r
\docType{data}
\name{theropod_data}
\alias{theropod_data}
\title{Theropod morphological data}
\format{
list
}
\usage{
data(theropod_data)
}
\description{
The morphological taxon-by-character matrix for theropod dinosaurs from Xu et al. 2023.
}
\examples{
data(theropod_data)
head(theropod_data)
}
\references{
Xu, L., Wang, M., Chen, R., Dong, L., Lin, M., Xu, X., Tang, J., You, H., Zhou, G., Wang, L. and He, W., 2023. A new avialan theropod from an emerging Jurassic terrestrial fauna. Nature, 621(7978), pp.336-343.
}
\keyword{datasets}
