% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeachingMaterials.r
\docType{data}
\name{theropod_tree}
\alias{theropod_tree}
\title{Theropod phylogenetic tree}
\format{
phylo
}
\usage{
data(theropod_tree)
}
\description{
A Maximum Clade Credibility tree of theropods modified from Xu et al. 2023.
}
\examples{
data(theropod_tree)
plot(theropod_tree, cex = 0.5)
}
\references{
Xu, L., Wang, M., Chen, R., Dong, L., Lin, M., Xu, X., Tang, J., You, H., Zhou, G., Wang, L. and He, W., 2023. A new avialan theropod from an emerging Jurassic terrestrial fauna. Nature, 621(7978), pp.336-343.
}
\keyword{datasets}
