% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeachingMaterials.r
\docType{data}
\name{sim_occ}
\alias{sim_occ}
\title{Simulated occurrence dataset}
\format{
data.frame
}
\usage{
data(sim_occ)
}
\description{
A simulated dataset of fossil occurrences analagous to data downloaded from the PBDB.
}
\examples{
data(sim_occ)
head(sim_occ)
}
\references{
Kocsis, Ádám T., Carl J. Reddin, John Alroy, and Wolfgang Kiessling. "The R package divDyn for quantifying diversity dynamics using fossil sampling data." Methods in Ecology and Evolution 10, no. 5 (2019): 735-743.
}
\keyword{datasets}
